Made the fuse interface only send reports when working outside of the .dupfs_sync directory.
Updated the ocmmit hook to read a config file to send out information.
It also now reads the revision and path and uses them.  The path arg
is used as the base for the config file, the revision is passed through

Now I need to test it, and ensure that the clients do the correct check
and then run a minimalistic svn update command